requirements.txt
# Core
asyncio>=3.4.3
aiohttp>=3.8.0
pydantic>=1.10.0
python-dotenv>=0.19.0

# Conductor
conductor-python>=3.9.0

# Logging & Monitoring
structlog>=22.0.0
prometheus-client>=0.15.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0

# Quality
black>=22.0.0
flake8>=5.0.0
mypy>=0.990
